+++ packages/tools/utils/src/array/fn/number/number.js
@@ -1,0 +1,32 @@
+/**
+* @module @svizzle/utils/array/fn/number/number
+*/
+
+import * as _ from "lamb";
+
+/**
+ * Return a function that computes the polynomial of the input number using the provided cofficients.
+ *
+ * @function
+ * @arg {array} cofficients - Array of cofficients, the exponent corresponding to the their index
+ * @return {function}
+ *
+ * @example
+const poly = makePolynomial([0,2,0,4]) // x => 2 x + 4 x^3
+poly(2) // 36
+poly(5) // 510
+ *
+ * @version 0.3.0
+ */
+export const makePolynomial = _.pipe([
+  _.reduceWith((acc, coefficient, exponent) => {
+    if (coefficient) {
+      acc.push(x => coefficient * Math.pow(x, exponent))
+    }
+    return acc;
+  }, []),
+  terms => _.pipe([
+    _.collect(terms),
+    _.reduceWith(_.sum)
+  ])
+]);